backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
fugashi[unidic]>=1.3.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
azure-cognitiveservices-speech>=1.35.0
praat-parselmouth>=0.4.0
fastdtw>=0.3.0
scipy>=1.10.0
numpy>=1.24.0
redis>=5.0.0
boto3>=1.34.0
